add custom stor to plug store and tag app, last db migrate

+++ likes/models.py
@@ -1,13 +1,11 @@
-from django.contrib.contenttypes.models import ContentType
-from django.contrib.contenttypes.fields import GenericForeignKey
 from django.db import models
 from django.contrib.auth.models import User
+from django.contrib.contenttypes.models import ContentType
+from django.contrib.contenttypes.fields import GenericForeignKey
+
 
 class LikedItem(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    # content_type shows the content the user has liked
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    # ref that particular object
-    object_id = models.PositiveBigIntegerField()
-    # showing that particular object
+    object_id = models.PositiveIntegerField()
     content_object = GenericForeignKey()
